bump to kui 5.4.1
Fixes #18

also part of #9 continued code cleanup
