+++ package.json
@@ -39,31 +39,31 @@
   ],
   "license": "Apache-2.0",
   "devDependencies": {
-    "@kui-shell/builder": "5.3.12",
-    "@kui-shell/core": "5.3.12",
-    "@kui-shell/plugin-bash-like": "5.3.12",
-    "@kui-shell/plugin-core-support": "5.3.12",
-    "@kui-shell/plugin-editor": "5.3.12",
-    "@kui-shell/plugin-proxy-support": "5.3.12",
-    "@kui-shell/proxy": "5.3.12",
-    "@kui-shell/test": "5.3.12",
-    "@kui-shell/webpack": "5.3.12",
+    "@kui-shell/builder": "5.4.1",
+    "@kui-shell/core": "5.4.1",
+    "@kui-shell/plugin-bash-like": "5.4.1",
+    "@kui-shell/plugin-core-support": "5.4.1",
+    "@kui-shell/plugin-editor": "5.4.1",
+    "@kui-shell/plugin-proxy-support": "5.4.1",
+    "@kui-shell/proxy": "5.4.1",
+    "@kui-shell/test": "5.4.1",
+    "@kui-shell/webpack": "5.4.1",
     "@types/mocha": "5.2.7",
-    "@types/node": "12.12.8",
+    "@types/node": "12.12.12",
     "@types/swagger-schema-official": "^2.0.20",
     "@types/uuid": "3.4.6",
     "@types/webdriverio": "4.13.3",
-    "@typescript-eslint/eslint-plugin": "2.7.0",
-    "@typescript-eslint/parser": "2.7.0",
+    "@typescript-eslint/eslint-plugin": "2.8.0",
+    "@typescript-eslint/parser": "2.8.0",
     "electron": "6.0.12",
-    "eslint": "6.6.0",
-    "eslint-config-prettier": "6.6.0",
+    "eslint": "6.7.0",
+    "eslint-config-prettier": "6.7.0",
     "eslint-config-standard": "14.1.0",
     "eslint-plugin-import": "2.18.2",
     "eslint-plugin-node": "10.0.0",
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-standard": "4.0.1",
-    "husky": "3.0.9",
+    "husky": "3.1.0",
     "lint-staged": "9.4.3",
     "mocha": "6.2.2",
     "prettier": "1.19.1",
