+++ plugins/plugin-grid/package.json
@@ -18,7 +18,7 @@
   ],
   "main": "dist/index.js",
   "dependencies": {
-    "pretty-ms": "5.0.0",
+    "pretty-ms": "5.1.0",
     "uuid": "3.3.3"
   },
   "kui": {
