+++ plugins/plugin-openwhisk-editor-extensions/package.json
@@ -27,7 +27,7 @@
   },
   "gitHead": "89de9f78e8a1a2bdd29d2e17d7c608ab006c6d32",
   "dependencies": {
-    "@kui-shell/plugin-editor": "^5.3.12",
+    "@kui-shell/plugin-editor": "^5.4.1",
     "debug": "4.1.1"
   }
 }
